+++ src/runtime_src/core/tools/common/XBUtilities.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2019-2020 Xilinx, Inc
+ * Copyright (C) 2019-2021 Xilinx, Inc
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may
  * not use this file except in compliance with the License. A copy of the
@@ -20,6 +20,7 @@
 // Include files
 // Please keep these to the bare minimum
 #include "core/common/device.h"
+#include "core/common/query_requests.h"
 
 #include <string>
 #include <memory>
@@ -136,6 +137,33 @@ namespace XBUtilities {
 
   std::string 
   parse_clock_id(const std::string& id);
+
+ /*
+  * xclbin locking
+  */
+  struct xclbin_lock
+  {
+    xclDeviceHandle m_handle;
+    xuid_t m_uuid;
+
+    xclbin_lock(std::shared_ptr<xrt_core::device> _dev)
+      : m_handle(_dev->get_device_handle())
+    {
+      auto xclbinid = xrt_core::device_query<xrt_core::query::xclbin_uuid>(_dev);
+
+      uuid_parse(xclbinid.c_str(), m_uuid);
+
+      if (uuid_is_null(m_uuid))
+        throw std::runtime_error("'uuid' invalid, please re-program xclbin.");
+
+      if (xclOpenContext(m_handle, m_uuid, std::numeric_limits<unsigned int>::max(), true))
+        throw std::runtime_error("'Failed to lock down xclbin");
+    }
+
+    ~xclbin_lock(){
+      xclCloseContext(m_handle, m_uuid, std::numeric_limits<unsigned int>::max());
+    }
+  };
 
 };
