+++ src/runtime_src/core/tools/xbutil2/SubCmdValidate.cpp
@@ -61,33 +61,6 @@ enum class test_status
 };
 
 /*
- * xclbin locking
- */
-struct xclbin_lock
-{
-  xclDeviceHandle m_handle;
-  xuid_t m_uuid;
-
-  xclbin_lock(std::shared_ptr<xrt_core::device> _dev)
-    : m_handle(_dev->get_device_handle())
-  {
-    auto xclbinid = xrt_core::device_query<xrt_core::query::xclbin_uuid>(_dev);
-
-    uuid_parse(xclbinid.c_str(), m_uuid);
-
-    if (uuid_is_null(m_uuid))
-      throw std::runtime_error("'uuid' invalid, please re-program xclbin.");
-
-    if (xclOpenContext(m_handle, m_uuid, std::numeric_limits<unsigned int>::max(), true))
-      throw std::runtime_error("'Failed to lock down xclbin");
-  }
-
-  ~xclbin_lock(){
-    xclCloseContext(m_handle, m_uuid, std::numeric_limits<unsigned int>::max());
-  }
-};
-
-/*
  * mini logger to log errors, warnings and details produced by the test cases
  */
 void logger(boost::property_tree::ptree& _ptTest, const std::string& tag, const std::string& msg)
@@ -900,7 +873,7 @@ p2pTest(const std::shared_ptr<xrt_core::device>& _dev, boost::property_tree::ptr
   }
 
   std::string msg;
-  xclbin_lock xclbin_lock(_dev);
+  XBU::xclbin_lock xclbin_lock(_dev);
   XBU::check_p2p_config(_dev.get(), msg);
 
   if(msg.find("Error") == 0) {
@@ -951,7 +924,7 @@ m2mTest(const std::shared_ptr<xrt_core::device>& _dev, boost::property_tree::ptr
     return;
   }
 
-  xclbin_lock xclbin_lock(_dev);
+  XBU::xclbin_lock xclbin_lock(_dev);
   uint32_t m2m_enabled = xrt_core::device_query<xrt_core::query::kds_numcdmas>(_dev);
   std::string name = xrt_core::device_query<xrt_core::query::rom_vbnv>(_dev);
 
@@ -1042,7 +1015,7 @@ bistTest(const std::shared_ptr<xrt_core::device>& _dev, boost::property_tree::pt
     return;
   }
 
-  xclbin_lock xclbin_lock(_dev);
+  XBU::xclbin_lock xclbin_lock(_dev);
 
   if (!clock_calibration(_dev, _dev->get_device_handle(), _ptTest))
      _ptTest.put("status", "failed");
